disallow type names in type parameters
The intent was that you could define a method like:

	Int Vec [ foo ^Bar | ... ]

which only applied to a T Vec type where T=Int.
This type of leads to ambiguity in parenthesized parameter lists:

	(T Eq) Vec [ = T Vec& ^Bool | ... ]

is probably supposed to define a new type variable T with constraint Eq.
But when parsing, it could equally as well be a single TypeName
for a type called Eq, which has an argument type T.

Instead, just forbid the short hand. You cna specialize to an Int Vec like so:

	(T Int) Vec [ foo ^Bar | ... ]

I suspect that such things will be rare anyway.