support multi-file module parsing
The start and end offsets are the byte offset into the text
as if all files were concatenated before parsing.
Each *File tracks its start offset into this range
along with the offset of its newlines, so file:line locations can be recovered
from just the simple integer offsets.

The root module name is an argument to the parser.
Each module path in the AST begins with the root module name.
Paths referring to identifiers in imported modules will be re-written by Check.