regenerate pea/grammar.go with sorted label declarations
There was an upstream fix in Peggy to generated deterministic output by sorting the label declarations. This change is a no-op, except that pea/grammar.go now has sorted label declarations. It should now be much easier to see true diffs in the parser source when the grammar is changed.