simplify the AST for types and add alias types
Instead of having Struct, Enum, and Virt, just have a single node called Type.
Instead of calling the types struct, enum, and virt, they are "and", "or", and "virtual".
Aliases are alternative names for existing types; their grammar is like assignment (:=).
Built-in types will use the Type node, but will needn't set the Alias, Fields, Cases, or Virts fields.