handle imports in Check
To do this, we now also represent modules in a tree instead of a flat map of paths. This lets us represent the location of a def based on how it is imported even if the ModPath of the def itself differs from the import path. So, each def has a single ModPath which is based on its path in its definining module, but it may be imported into mulitple other modules and referenced by any of those names.