add some checking of type and method receiver signatures
For type signatures, just check that the constraints exist, but not that they are met.
For receiver signatures, check that the type exists, that it is indeed a type, that its constraints exist, and that the parameter counts match.

This will all likely shuffle around once we have proper checking of parameterized defs.