check imported modules
We need to call check on imported modules (including built-in),
because later code will assume that the ASTs for the modules' definitions
are fully linked with TypeName.Type set and so on.

In addition, the built-in module definitions were needlessly duplicated
into the currently checked module in the module tree. Stop doing this.
This is not needed, because definition lookup checks all containing modules,
and the built-in module is always at the top of the containment hierarchy.