make typeStringForUser return a string, not build one
Originally I had no intent for this function to see heavy use, but for better or worse, it's very heavily used; and it was much too hard to use since it required all callers to create a string builder. This simplifies it.

I still think this funciton should be somehow refactored. Probably the .String() methods should all be re-done to return user-friendly string versions. We can add DebugString() or some such which returns a more noisy version.