move checkLiftedMsg and friends into inst.go
This method could sort of go in either inst.go or check.go, since instantiating and checking go hand-in-hand here. But check.go is getting long, so let's move this code to the other file for now. Now all functions that populate Mod.Insts are in the same place at least: inst.go.