add Sub Def to the AST to replace ModPath in definitions
Instead of flattening the module tree and tracking definition modules using ModPath, keep the tree structure, represented by the Sub Def node. ModPath is now just for TypeNames and Call receivers.

The reason for this change is because we need the tree structure on the type checking side anyway, so it simplifies a lot to keep it in the AST too.