rename Var to Val and Parm to Var
Some uses of what was formerly Parm are better described as Var(iable)s. This will be particularly true in the type checker, for example Local variables are not a parameter to anything. In generall, this type was used to describe what are effectively named memory locations. That's a Var.

To free up the name Var, I renamed the module-level variable thingy to Val. Really, these won't be able to vary anyway, so it's more of a Val(ue) than a Var(iable).