revamp the grammar of definitions
Remove the concept of sub-modules; change ModPath to just a single ModName. When importing a module, all of its symbols will need to be referenced by prefixing with the #modname. Type aliases allow aliasing types into the current module namespace without requiring #modname before it.

Each definition starts with a keyword declaring the type of thing being defined. Whether the keyword is capital or lower-case dictates whether it is public or private.