unexport the Import type
It doesn't need to be exposed in the API. Importers can just return a []Def. This is better, because Import had a redundant Path field that was confusing to be returning from an Importer since the path was already the input, and it had an even more redundant/confusing Name field, which an Importer cannot ever set correctly, since it would need to be deterimened per-file (assuming we ever support re-naming imports).