nil-out imported Def.ast fields
A future importer should read the type tree for imports from a file, not parse them from source. In that case, the AST nodes will not be available. We need to make sure that we don't add any assumptions that the AST nodes of imported definitions exist. So, let's clear them out now to expose any such issues early.