make the argument to numeric type binary methods a reference
This is to make these play nicer with interfaces which will need to work on bigger types that are better to have a reference argument. For example,

type T Eq {[= T& ^Bool]}

can be a convenient constraint for types with an = method. Some times may be better passed by reference, and these types would presumably take & as the parameter. If the numeric types don't take a & parameter, they couldn't satisfy this interface.

(Maybe some day & won't count in type-checking interface parameters/return.)