disallow constructing unexported types
This can come up if an imported module has an exported alias to an unexported type:

	// foo.pea
	Type Foo := foo.
	type foo {x: int}

	// bar.pea
	import "foo"
	_ #foo Foo := [{x: 5}] // disallowed by this change

This pattern can be useful. It will be Pea's answer to private or unexported class or struct fields in other languages:

	// vec.pea
	Type T Vec := T vec

	type T vec {
		// These fields are not exposed to importing modules.
		data: T Array
		size: Int
	}

I also added tests to ensure that built-in methods of private virtual- and or-types are not accessable.