deal with cycles in debugString and String
Types can be cyclic. debugString needs to handle cycles so it doesn't hit infinite recursion. We do this with a map, and breaking the cycle by just printing the pointer so the user can refer back to other output to see the type.

String can avoid the cycles by simply not recurring on them. Instead of using TypeName.Type for TypeName strings, just print the TypeName contents itself. This means that String doesn't print the resolved type of an alias, but the alias itself. This may actually be desirable.