remove panicIf and add bugs() methods instead
The previous way of checking for bugs in the code was to use a newXyz method for each type Xyz, which was filed with panicIf assertions. The problem with this approach is that when a panic occurs (for example in a test), then it is extremely difficult to see what statement caused the panic — there is no test output.

Another, but less important, disadvantage is that a type constructed without the newXyz method would not get any checking.

The new approach in this change is to introduce a bug() method to all statements. It returns a string that is "" if there are no bugs, otherwise it is the description of something buggy in the statement. Bugs are output in the string representation of the statement, so they will show up as diffs in unit test output. This makes them much easier to see where a problem occurs.