change Arg to the argument value, not its address
Some arguments could be passed by register, which is not addressable. In LLVM, it seems like parameter/argument names are the parameter type, not the address of the parameter. So this change is inline with LLVM.

However, to support referencing a parameter, for non-pass-by-value parameters, they need to be explicitly copied into an allocated location in the function preamble. From then on, parameter accesses use that alloc instead of the Arg value. The optimization pass should be able to remove most of these needless allocs.