add a field to BlockType for containing function return slot capture
The basic package used to just access one-over the number of fields in Type.Fields for the return slot capture. This change makes it an explicit field generated by the types package instead.

The downstream code generation will sometimes only have the *types.Type of a BlockType, and it needs to know to emit a field for the function return capture. If there isn't an explicit field for it, then there is no way to know to emit it.

This also simplifies some code in basic, as it no longer needs to assume that fields will go one-beyond the return of types.Type.Fields.