allow assignment to self
The self identifier is now the actual receiver value, not an implicit reference parameter. This means that assigning to self changes the receiver's value, not a reference parameter value. This is more in line with how I would expect it behave. It also allows making methods that modify the receiver (as opposed to only methods that modify receiver fields). For example:

// increments the value of the receiver by one.
Meth Int [increment | self := self + 1]