make inlining less aggressive
Previously inlining would essentially inline everything unless there was a call to a function that is only declared (no body). This changes the inlining heuristic to be only leaf calls or calls to function with a Fun-type parameter. The latter is in order to allow further block literal inlining which can remove virtual calls.