add ast.ReadImports
It returns just the imports for a source file.

The current implementation is needlessly inefficient: it reads the entire source file, fully parses it, and returns the imports off of the AST. Since imports always come first, a more efficient implementation would read and parse just enough of the source file to get the imports and no more.