fix a bug where module variable init functions were not generated
There was a missing case in function name mangling: if f.Fun==nil but f.Val != nil, then it is the function initializing a single module variable. But this was getting conflated wit the whole module's initialization function, and consequently these were getting merged and dropped.

It didn't show up in tests, because after the opt pass, these were getting inlined.