add the test keyword
Tests are like functions, but they:
* cannot be called,
* cannot have parameters,
* cannot have a return value,
* and must have a body

The peac command has a new flag: -test. When set to true, the compiler builds a test binary instead of a normal binary. The test binary runs each test in serial. Tests fail if they panic, and the line of the test body that paniced is printed in the failure message.

The logic ignoring _test.pea files is now gone; tests can live in any file.