remove merged .go file after building the executable
It gets in the way having a .go file hanging around in the directory with my other go source, which is annoying. But also, traditional compilation has source files, object files, and executables. The source files here correspond to .pea files, the object files to .peago files, and the executable is the executable. The merged .go file doesn't really fit into this model; it can be thought of as more a detail of the last linking phase. So just make it temporary and remove it when done.