fix bugs with the String atByte method
The BUG check for Index was reporting an error if the receiver was String. Now it does not.

The addIndex function didn't handle String receivers at all, and set the Index type to a bogus value. Now it sets to UInt8.

The Go code for Index always took the address of the element. Now it only takes the address for non-String receivers.