refactor location handling to fix bugs with imported definitions
Motivation:

The previous assumption was that setting .AST=nil for imported (and built-in) defs in the type checker would be enough to flag all imported nodes as unlocatable. However, this is both wrong and undesirable.

It is wrong, because non-def nodes from imported modules are still pulled into the types tree, and these do not have their ASTs set to nil. These nodes are accessable by basic and other down-stream packages. If their locations are used, they were not accounted for in the primary module's AST locations. So, for example, when gengo emits the location of a panic: call, if that panic is inside the instantiation of an imported, parameterized function, the location is completely bogus.

It is undesirable because, we do want the source locations of imported nodes. Why not? These are used for error reporting by the type package during compilation and will be emitted into debug tables in the generated binary. In the first case, we can assume that the source is most often available (we aren't compiling from an object file someone emailed to us), and it's convenient to be able to click through to the error in another imported module. And in the second case, we certainly want to emit the source location infromation in debugging tables so that panics and stack traces have the correct location information.

What this change does:

This change adds a new package called loc that contains the location handling previously stuck in the ast package. The location tracking structures (loc.Files) are no longer hidden in the ast package, but are available to other packages too.

There are now two ways to create a parser: 1) you get default empty loc.Files, and 2) you pass in a *loc.Files to append to. Most callers will use #1. However, importers in the types package will use #2, passing the loc.Files of the primary module to be appended to on import. Finally, if a parser is created with a nil *loc.Files, then it sets all location information to -1. This mode is used for parsing the univ package.

When looking for a loc.Loc from a loc.Files, the return is now a pointer. If the requested offsets are negative (as they would be for the univ package) or beyond the end of all files (this would be a bug), then the result is nil. Since the type checker now always has non-nil AST pointers, checking this *loc.Loc's nil-ness is the new way to determine whether the Node's location can be printed.