fix a bug creating a unique name for instantiated funs
This fixes a bug where a module instantiated an imported fun with the same name as a fun defined in the instantiating module. Before this change, the instance had its modpath set to the instantiating module, and that caused its unique name to conflict with the in-module definition:

func T [new ^T Xyz |]
test [abc |
	// These two instantiated methods would get the same name.
	x Int Xyz := new.
	y Int Abc := #abc new.
]

Instead, set the ModPath of instances to their definiting module. Of course, we still need to differenciate instances that may have type parameters with method sets unique to the instantiating module. (Currently, we conservatively do this for anything with type constraints.) To support this, add a new InstModPath filed to Fun.