fix a language bug: case method block args must be values, not references
In general, it should not be possible to get a reference to a field of a case type. The reason is that the case type itself can be assigned to by a different case, which would invalidate the memory referenced.

This only comes up in case methods. The arguments previously were Funs that took as an argument a reference to the case value. This change makes it a value instead.