remove types.Call.typ
This is always supposed to be the same as call.Msgs[len(call.Msgs)-1].Type(), but has been the cause of various bugs when I forgot to actually set it. Instead, just remove the field and have Call.Type() return the correct value.

Of course this means that Msg.typ must altways be correctly set. So do that here too.