don't check the source directory when deciding to rebuild a module
Since the output file is written to the source directory, the source directory is always modified after the output file. This means that we effectively recompile from scratch every time. Instead, for the time being, don't check the source directory.

However, this is buggy in the face of source files being deleted. If a source file is deleted, the timestamp change isn't detected (there is no more file to check the timestamp of), so we won't pick up the change. Instead, we should write the list of source files to the .peago file so we can detect when one goes away.