allow numerical type inference for rune literals
Rune literals used to always be Int32. Now, we use the infer type to determine what the type is using the same algorithm as for int literals. This is useful, since it is common to use a rune literal as a UInt8.

I also slightly refactored the checkInt and checkFloat functions to use a switch statement and mirror the new checkRune logic, which I think is slightly easier to read.