search a type's defining module for tag-less method calls
This allows you to lower-case import a module defining a type but still access its methods without a mod tag.

This is useful because a very common case is to import a module for its type and corresponding methods. Those methods may be needed to implemented an interface. Without this change, the module would need to be capital-Imported, instead of lower-case.