fix a crash with chained alias definitions of differing arity
When instantiating an alias type, we used the alias targe TypeName to get the arguments for substitution. If that TypeName is itself an alias, the arity of its arguments may differ from those of the target type. We need to use the arguments of the target type itself, not the TypeName.