emit methods for virtual type virtual methods
Previously we didn't emit methods for virtual type virtual methods. These are handled by a call to an entry in the virtual table. However, this caused a compile failure when converting one virtual type to another. The conversion expects a static function to fill the slot in the virtual table. So, we emit such a function, which just wraps the virtual call on the inner type.

In the future, we should probably not need such a method. Instead, we can handle virtual-to-virtual conversion as a special case and use the entry from one virtual table to fill the entry in the other.