fix bug substituting function local variables
We were substituting a function instance's Locals field in subFun, which is called when a function instance is created. However, at that time, we may not have checked the definition's body yet, so it may not have its Locals field populated. When later substituting the function body, we will fail to find the locals.

Instead, don't populate the instance Locals field until substituting its body.